+++ packages/plugin-workflow-approval/src/client/common/components/FormBlock.provider.tsx
@@ -14,7 +14,7 @@ import {
 } from '@tachybase/client';
 import { createForm, RecursionField, useField, useFieldSchema } from '@tachybase/schema';
 
-import { useContextApprovalExecution } from '../../usage/pc/block/common/ApprovalExecution.provider';
+import { useContextApprovalExecution } from '../../usage/common/contexts/approvalExecution';
 import { useContextApprovalRecords } from '../../usage/pc/block/todos-table/providers/ApprovalExecutions.provider';
 
 export const FormBlockProvider = (props) => {
