+++ packages/plugin-workflow-approval/src/client/usage/pc/block/initiations-table/hooks/useFormBlockProps.tsx
@@ -3,9 +3,9 @@ import { useCurrentUserContext } from '@tachybase/client';
 import { useForm } from '@tachybase/schema';
 
 import { approvalInitiationStatusMap } from '../../../../../common/constants/approval-initiation-status-options';
+import { useContextApprovalExecution } from '../../../../common/contexts/approvalExecution';
 import { useApproval } from '../../../common/ApprovalData.provider';
 import { useResubmit } from '../../../common/Resubmit.provider';
-import { useContextApprovalExecution } from '../../common/ApprovalExecution.provider';
 
 export function useFormBlockProps() {
   const approval = useApproval() as any;
